feat: change types

+++ Sources/SolanaSwift/Programs/TokenPrograms/Token2022Program/ParsedTransferFeeConfig.swift
@@ -4,12 +4,12 @@ public struct ParsedTransferFeeConfig: Codable {
     public static let name: String = "transferFeeConfig"
 
     public struct TransferFee: Codable {
-        let epoch: Int64
-        let maximumFee: Double
-        let transferFeeBasisPoints: Int64
+        let epoch: UInt64
+        let maximumFee: UInt64
+        let transferFeeBasisPoints: UInt64
     }
 
     public let newerTransferFee, olderTransferFee: TransferFee
     public let transferFeeConfigAuthority, withdrawWithheldAuthority: String
-    public let withheldAmount: Int64
+    public let withheldAmount: UInt64
 }
